[SIX-247] fix : 같은 빈을 공유하는 통합 테스트로 인한 오류 해결
