+++ onedayhero-application/src/test/java/com/sixheroes/onedayheroapplication/mission/MissionServiceTest.java
@@ -3,8 +3,6 @@ package com.sixheroes.onedayheroapplication.mission;
 import com.sixheroes.onedayheroapplication.IntegrationApplicationTest;
 import com.sixheroes.onedayheroapplication.global.s3.dto.request.S3ImageUploadServiceRequest;
 import com.sixheroes.onedayheroapplication.global.s3.dto.response.S3ImageUploadServiceResponse;
-import com.sixheroes.onedayheroapplication.main.request.UserPositionServiceRequest;
-import com.sixheroes.onedayheroapplication.main.response.MissionSoonExpiredResponse;
 import com.sixheroes.onedayheroapplication.mission.request.MissionCreateServiceRequest;
 import com.sixheroes.onedayheroapplication.mission.request.MissionExtendServiceRequest;
 import com.sixheroes.onedayheroapplication.mission.request.MissionInfoServiceRequest;
@@ -13,13 +11,11 @@ import com.sixheroes.onedayheroapplication.mission.response.MissionCategoryRespo
 import com.sixheroes.onedayheroapplication.region.response.RegionResponse;
 import com.sixheroes.onedayherocommon.error.ErrorCode;
 import com.sixheroes.onedayherodomain.mission.*;
-import com.sixheroes.onedayherodomain.mission.repository.response.MainMissionQueryResponse;
 import org.apache.http.entity.ContentType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
-import org.locationtech.jts.geom.Point;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -935,40 +931,6 @@ class MissionServiceTest extends IntegrationApplicationTest {
                 );
     }
 
-    @DisplayName("메인 페이지를 조회 할 수 있다.")
-    @Test
-    void findMainResponse() {
-        // given
-        var latitude = 37.37898914571587;
-        var longitude = 121.66272838210327;
-        var userPosition = UserPositionServiceRequest.builder()
-                .latitude(latitude)
-                .longitude(longitude)
-                .build();
-        var userId = 1L;
-        var serverTime = LocalDateTime.of(LocalDate.of(2023, 11, 23),
-                LocalTime.MIDNIGHT);
-
-        var missionCategoryResponse = MissionCategoryResponse.builder()
-                .id(1L)
-                .code("MC_001")
-                .name("서빙")
-                .build();
-
-        var queryResponse = List.of((MainMissionQueryResponse) new MainMissionQueryTestDto());
-
-        given(missionNativeQueryRepository.findSoonExpiredMissionByLocation(any(String.class), any(Integer.class), any(Long.class), any(LocalDateTime.class), any(LocalDateTime.class)))
-                .willReturn(queryResponse);
-
-        var result = MissionSoonExpiredResponse.from(queryResponse.get(0), "s3::/path");
-
-        // when
-        var mainResponse = mainService.findMainResponse(userId, userPosition, serverTime);
-
-        // then
-        assertThat(mainResponse.missionCategories().get(0)).isEqualTo(missionCategoryResponse);
-        assertThat(mainResponse.soonExpiredMissions().get(0)).isEqualTo(result);
-    }
 
     private MissionCreateServiceRequest createMissionCreateServiceRequest(
             MissionInfoServiceRequest missionInfoServiceRequest
@@ -1087,103 +1049,5 @@ class MissionServiceTest extends IntegrationApplicationTest {
                 .price(10000)
                 .serverTime(serverTime)
                 .build();
-    }
-
-    private static class MainMissionQueryTestDto implements MainMissionQueryResponse {
-        @Override
-        public Long getId() {
-            return 1L;
-        }
-
-        @Override
-        public Long getCategoryId() {
-            return 1L;
-        }
-
-        @Override
-        public String getCategoryCode() {
-            return "MC_001";
-        }
-
-        @Override
-        public String getCategoryName() {
-            return "서빙";
-        }
-
-        @Override
-        public Long getCitizenId() {
-            return 1L;
-        }
-
-        @Override
-        public Long getRegionId() {
-            return 1L;
-        }
-
-        @Override
-        public String getSi() {
-            return "서울시";
-        }
-
-        @Override
-        public String getGu() {
-            return "강남구";
-        }
-
-        @Override
-        public String getDong() {
-            return "역삼동";
-        }
-
-        @Override
-        public Point getLocation() {
-            return Mission.createPoint(121.66272838210327, 37.37898914571587);
-        }
-
-        @Override
-        public String getTitle() {
-            return "미션 제목";
-        }
-
-        @Override
-        public LocalDate getMissionDate() {
-            return LocalDate.of(2022, 11, 23);
-        }
-
-        @Override
-        public LocalTime getStartTime() {
-            return LocalTime.of(12, 30);
-        }
-
-        @Override
-        public LocalTime getEndTime() {
-            return LocalTime.of(13, 0);
-        }
-
-        @Override
-        public LocalDateTime getDeadlineTime() {
-            return LocalDateTime.of(LocalDate.of(2022, 11, 23),
-                    LocalTime.of(12, 30));
-        }
-
-        @Override
-        public Integer getPrice() {
-            return 4000;
-        }
-
-        @Override
-        public Integer getBookmarkCount() {
-            return 3;
-        }
-
-        @Override
-        public MissionStatus getMissionStatus() {
-            return MissionStatus.MATCHING;
-        }
-
-        @Override
-        public Long getBookmarkId() {
-            return 1L;
-        }
     }
 }
